@@ -5,7 +5,7 @@ description = ""
 authors = ["Nils Olsson <nilso@enosis.net>"]
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.10"
 python-dotenv = {version = "^0.20.0", optional = true}
 
 [tool.poetry.extras]
@@ -19,12 +19,15 @@ isort = "^5.10.1"
 pyflakes = "2.4.0"
 rich = "^12.5.1"
 colorama = "^0.4.5"
+pyright = "^1.1.267"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.pyright]
+venvPath = "."
+venv = ".venv"
 typeCheckingMode = "strict"
 reportImplicitStringConcatenation = true
 reportCallInDefaultInitializer = true
